Boost stuff, might cause an error
I tried building on FreeBSD and got some errors, so I did some
searching. Apparently I need to add -DBOOST_NO_CXX11_SCOPED_ENUMS to
the compiler to build with clang, so I added that with add_definitions
with CMake, and I'm not sure if that's the correct way to do it. I
tested it and it stills works fine on Gentoo and builds on FreeBSD now,
so I hope it was the correct thing to do.