Completely changed command system
ConfigFileReader now stores a new object called and environment. This
environment stores the options, and may contain config file variables
in the future. Now, to create an action, one must also have an
environment to get it from. This meant that I had to change a lot of
declarations and implementations.