Started writing new config file stuff
Renamed filecheck to filecheckaction and made it extend the
ModuleAction class.

Rewrote the module parser to have install, uninstall, and update
states, and a preliminary file install section.

Also added a function called verboseMessage to the config file reader,
which prints info about the current line which cleaned up a lot of my
code.